(fix) Statistic
